# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "DataSelectionResponse": {
        "additionalProperties": false,
        "properties": {
          "session_ids": {
            "items": {
              "type": "string"
            },
            "title": "Session Ids",
            "type": "array"
          },
          "since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "required": [
          "since",
          "until",
          "timezone",
          "session_ids"
        ],
        "title": "DataSelectionResponse",
        "type": "object"
      },
      "DismissRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "DismissRequest",
        "type": "object"
      },
      "DraftRequest": {
        "additionalProperties": false,
        "properties": {
          "contents": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Contents",
            "type": "object"
          },
          "expected_draft_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Draft Revision"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "contents"
        ],
        "title": "DraftRequest",
        "type": "object"
      },
      "DraftResetRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_draft_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Draft Revision"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "DraftResetRequest",
        "type": "object"
      },
...
      "EffectiveRunConfig": {
        "additionalProperties": false,
        "properties": {
          "candidate_budget": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Candidate Budget",
            "type": "integer"
          },
          "force": {
            "title": "Force",
            "type": "boolean"
          },
          "judging_context": {
            "$ref": "#/components/schemas/JudgingContextPolicy"
          },
          "model_catalog_version": {
            "title": "Model Catalog Version",
            "type": "string"
          },
          "models": {
            "$ref": "#/components/schemas/EffectiveModelSelection"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "rubric_catalog_version": {
            "title": "Rubric Catalog Version",
            "type": "string"
          },
          "rubrics": {
            "items": {
              "$ref": "#/components/schemas/RubricDescriptor"
            },
            "title": "Rubrics",
            "type": "array"
          },
          "schema_version": {
            "const": "5",
            "default": "5",
            "title": "Schema Version",
            "type": "string"
          },
          "selection_warnings": {
            "items": {
              "$ref": "#/components/schemas/SelectionWarning"
            },
            "title": "Selection Warnings",
            "type": "array"
          }
        },
        "required": [
          "pipeline_version",
          "model_catalog_version",
          "rubric_catalog_version",
          "models",
          "rubrics",
          "selection_warnings",
          "judging_context",
          "candidate_budget",
          "force"
        ],
        "title": "EffectiveRunConfig",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JudgingPlanResponse": {
        "additionalProperties": false,
        "properties": {
          "cohort_id": {
            "title": "Cohort Id",
            "type": "string"
          },
          "input_policy": {
            "$ref": "#/components/schemas/JudgingContextPolicy"
          },
          "panel_size": {
            "title": "Panel Size",
            "type": "integer"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "protocol": {
            "additionalProperties": true,
            "title": "Protocol",
            "type": "object"
          },
          "requested_rubrics": {
            "items": {
              "$ref": "#/components/schemas/RubricDescriptor"
            },
            "title": "Requested Rubrics",
            "type": "array"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/JudgingPlanSessionResponse"
            },
            "title": "Sessions",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/JudgingPlanTotalsResponse"
          }
        },
        "required": [
          "plan_id",
          "schema_version",
          "cohort_id",
          "panel_size",
          "requested_rubrics",
          "input_policy",
          "protocol",
          "totals",
          "sessions"
        ],
        "title": "JudgingPlanResponse",
        "type": "object"
      },
...
      "JudgingProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt_summaries": {
            "items": {
              "$ref": "#/components/schemas/JudgingAttemptSummaryResponse"
            },
            "title": "Attempt Summaries",
            "type": "array"
          },
          "attempt_summaries_truncated": {
            "title": "Attempt Summaries Truncated",
            "type": "boolean"
          },
          "attempt_summary_count": {
            "title": "Attempt Summary Count",
            "type": "integer"
          },
          "coverage_complete": {
            "title": "Coverage Complete",
            "type": "boolean"
          },
          "digest_steps_completed": {
            "title": "Digest Steps Completed",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/JudgingActivityEventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "failure_count": {
            "title": "Failure Count",
            "type": "integer"
          },
          "failure_detail_count": {
            "title": "Failure Detail Count",
            "type": "integer"
          },
          "failure_details": {
            "items": {
              "$ref": "#/components/schemas/JudgingFailureDetailResponse"
            },
            "title": "Failure Details",
            "type": "array"
          },
          "failure_details_truncated": {
            "title": "Failure Details Truncated",
            "type": "boolean"
          },
          "maximum_digest_steps": {
            "title": "Maximum Digest Steps",
            "type": "integer"
          },
          "maximum_merge_steps": {
            "title": "Maximum Merge Steps",
            "type": "integer"
          },
          "maximum_reviewer_attempts": {
            "title": "Maximum Reviewer Attempts",
            "type": "integer"
          },
          "maximum_window_steps": {
            "title": "Maximum Window Steps",
            "type": "integer"
          },
          "merge_steps_completed": {
            "title": "Merge Steps Completed",
            "type": "integer"
          },
          "minimum_reviewer_attempts": {
            "title": "Minimum Reviewer Attempts",
            "type": "integer"
          },
          "not_evaluable_rubrics": {
            "title": "Not Evaluable Rubrics",
            "type": "integer"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "planned_rubrics": {
            "title": "Planned Rubrics",
            "type": "integer"
          },
          "rated_rubrics": {
            "title": "Rated Rubrics",
            "type": "integer"
          },
          "reviewer_attempts_completed": {
            "title": "Reviewer Attempts Completed",
            "type": "integer"
          },
          "rubrics_completed": {
            "title": "Rubrics Completed",
            "type": "integer"
          },
          "scores_written": {
            "title": "Scores Written",
            "type": "integer"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          },
          "window_steps_completed": {
            "title": "Window Steps Completed",
            "type": "integer"
          },
          "write_failure_count": {
            "title": "Write Failure Count",
            "type": "integer"
          }
        },
        "required": [
          "plan_id",
          "planned_rubrics",
          "rubrics_completed",
          "rated_rubrics",
          "not_evaluable_rubrics",
          "minimum_reviewer_attempts",
          "maximum_reviewer_attempts",
          "reviewer_attempts_completed",
          "digest_steps_completed",
          "maximum_digest_steps",
          "window_steps_completed",
          "maximum_window_steps",
          "merge_steps_completed",
          "maximum_merge_steps",
          "scores_written",
          "failure_count",
          "write_failure_count",
          "coverage_complete",
          "phase",
          "status_message",
          "started_at",
          "events",
          "attempt_summary_count",
          "attempt_summaries_truncated",
          "attempt_summaries",
          "failure_detail_count",
          "failure_details_truncated",
          "failure_details"
        ],
        "title": "JudgingProgressResponse",
        "type": "object"
      },
...
      "PromoteRequest": {
        "additionalProperties": false,
        "properties": {
          "acknowledge_unevaluated": {
            "default": false,
            "title": "Acknowledge Unevaluated",
            "type": "boolean"
          },
          "expected_draft_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Draft Revision"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "idempotency_key"
        ],
        "title": "PromoteRequest",
        "type": "object"
      },
...
      "ReflectionResultResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/SuccessfulReflectionResultResponse"
          },
          {
            "$ref": "#/components/schemas/EmptyReflectionResultResponse"
          }
        ],
        "title": "ReflectionResultResponse"
      },
...
      "RunListResponse": {
        "additionalProperties": false,
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummaryResponse"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "title": "RunListResponse",
        "type": "object"
      },
      "RunResponse": {
        "additionalProperties": false,
        "properties": {
          "auto_run": {
            "title": "Auto Run",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_stage_succeeded": {
            "title": "Current Stage Succeeded",
            "type": "boolean"
          },
          "data_selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataSelectionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffectiveRunConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "judging_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JudgingPlanResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "judging_progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JudgingProgressResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "judging_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JudgingProgressResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reflecting_progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectingProgressResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reflecting_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectionResultResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reflection_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectionReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reflection_review_revision": {
            "title": "Reflection Review Revision",
            "type": "integer"
          },
          "run_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scoring_progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoringProgressResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoring_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoringResultResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "created",
              "scoring",
              "judging",
              "reflecting",
              "complete",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "turn_cohort": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Cohort"
          }
        },
        "required": [
          "run_id",
          "status",
          "current_stage_succeeded",
          "created_at",
          "auto_run",
          "data_selection",
          "run_config",
          "effective_config",
          "turn_cohort",
          "judging_plan",
          "scoring_progress",
          "scoring_result",
          "judging_progress",
          "judging_result",
          "reflecting_progress",
          "reflecting_result",
          "reflection_review",
          "reflection_review_revision",
          "error"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "RunSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "current_stage_succeeded": {
            "title": "Current Stage Succeeded",
            "type": "boolean"
          },
          "review_state": {
            "enum": [
              "none",
              "review-needed",
              "promoted",
              "partial",
              "dismissed",
              "no-change",
              "no-valid-proposal"
            ],
            "title": "Review State",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunSelectionSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "created",
              "scoring",
              "judging",
              "reflecting",
              "complete",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "current_stage_succeeded",
          "created_at",
          "selection",
          "review_state"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
...
      "ScoringProgressResponse": {
        "additionalProperties": false,
        "properties": {
          "scored": {
            "title": "Scored",
            "type": "integer"
          },
          "status_message": {
            "title": "Status Message",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "turn_details": {
            "items": {
              "$ref": "#/components/schemas/ScoringTurnDetailResponse"
            },
            "title": "Turn Details",
            "type": "array"
          },
          "written": {
            "title": "Written",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "scored",
          "written",
          "status_message",
          "turn_details"
        ],
        "title": "ScoringProgressResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "weave_agent_signals__routes__reviews__SelectionRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Candidate Id",
            "type": "string"
          },
          "discard_draft": {
            "default": false,
            "title": "Discard Draft",
            "type": "boolean"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "candidate_id",
          "expected_revision"
        ],
        "title": "SelectionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "weave-agent-signals",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_api_runs_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/dismiss": {
      "post": {
        "operationId": "dismiss_api_runs__run_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/runs/{run_id}/promote": {
      "post": {
        "operationId": "promote_api_runs__run_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/runs/{run_id}/reflection_draft": {
      "delete": {
        "operationId": "reset_draft_api_runs__run_id__reflection_draft_delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Draft",
        "tags": [
          "reviews"
        ]
      },
      "put": {
        "operationId": "save_draft_api_runs__run_id__reflection_draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Draft",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/runs/{run_id}/reflection_selection": {
      "put": {
        "operationId": "select_candidate_api_runs__run_id__reflection_selection_put",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/weave_agent_signals__routes__reviews__SelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Candidate",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}